refactor(game): LottoGame과 InputParser를 classicLotto로 이동
완전한 패키지 분리를 위해 LottoGame과 InputParser 이동
- LottoGame, InputParser 패키지 변경
- Application import 경로 수정
- 모든 로또 관련 코드가 game.classicLotto 패키지에 통합
