+++ src/main/java/lotto/Application.java
@@ -2,6 +2,7 @@ package lotto;
 
 import lotto.engine.GameEngine;
 import lotto.game.Game;
+import lotto.game.classicLotto.LottoGame;
 
 public class Application {
     public static void main(String[] args) {
